+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * feature#259 - improve all panels
 
+* feature: Limit database check to size 1GB
+
 * New event highlighting system
 
 --- 4.0.3 ---